+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Upgrade to Gradle 7.5.1
 - Declare Gradle Plugin functional test as JVM test suite
 
 ## [0.5.0] - 2022-07-21